requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic>=0.50.0
supabase==2.10.0
pandas==2.2.3
numpy==2.2.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
structlog==24.4.0
tavily-python==0.5.0
cachetools==5.5.2
rapidfuzz>=3.0
ruff==0.12.9
